+++ src/interchange/src/protobuf/decode.rs
@@ -9,7 +9,6 @@
 
 use anyhow::{anyhow, bail, Context, Result};
 
-use num_traits::ToPrimitive;
 use ordered_float::OrderedFloat;
 use serde::de::Deserialize;
 use serde_protobuf::de::Deserializer;
@@ -18,9 +17,9 @@ use serde_protobuf::value::Value as ProtoValue;
 use serde_value::Value as SerdeValue;
 
 use repr::adt::numeric::Numeric;
-use repr::{Datum, DatumList, Row};
+use repr::{ColumnType, Datum, DatumList, Row, ScalarType};
 
-use crate::protobuf::proto_message_name;
+use crate::protobuf::{proto_message_name, validate_descriptors};
 
 /// Manages required metadata to read protobuf
 #[derive(Debug)]
@@ -56,6 +55,7 @@ impl Decoder {
                 .map_err(|e| anyhow!("Creating an input stream to parse protobuf: {}", e))?;
         let deserialized_message =
             SerdeValue::deserialize(&mut deserializer).context("Deserializing into rust object")?;
+        let relation_type = validate_descriptors(&self.message_name, &self.descriptors)?;
 
         let msg_name = &self.message_name;
         let mut packer = &mut self.packer;
@@ -68,6 +68,7 @@ impl Decoder {
                     msg_name
                 )
             })?,
+            &relation_type.typ().column_types,
             &mut packer,
         )?;
         if push_metadata {
@@ -93,6 +94,7 @@ fn extract_row_into(
     deserialized_message: SerdeValue,
     descriptors: &Descriptors,
     message_descriptors: &MessageDescriptor,
+    column_types: &[ColumnType],
     packer: &mut Row,
 ) -> Result<()> {
     let deserialized_message = match deserialized_message {
@@ -101,12 +103,11 @@ fn extract_row_into(
     };
 
     // TODO: This is actually unpacking a row, it should always return json
-    for f in message_descriptors.fields().iter() {
+    for (f, column_type) in message_descriptors.fields().iter().zip(column_types) {
         let key = SerdeValue::String(f.name().to_string());
         let value = deserialized_message.get(&key);
-
         if let Some(value) = value {
-            json_from_serde_value(&value, packer, f, descriptors)?;
+            json_from_serde_value(&value, packer, f, descriptors, column_type)?;
         } else {
             packer.push(default_datum_from_field(f, descriptors)?);
         }
@@ -124,6 +125,7 @@ fn json_from_serde_value(
     packer: &mut Row,
     f: &FieldDescriptor,
     descriptors: &Descriptors,
+    column_type: &ColumnType,
 ) -> Result<()> {
     packer.push(match val {
         SerdeValue::Bool(true) => Datum::True,
@@ -142,13 +144,13 @@ fn json_from_serde_value(
         SerdeValue::Bytes(b) => Datum::Bytes(b),
         SerdeValue::Option(s) => {
             if let Some(s) = s {
-                return json_from_serde_value(&s, packer, f, descriptors);
+                return json_from_serde_value(&s, packer, f, descriptors, column_type);
             }
 
             default_datum_from_field(f, descriptors)?
         }
         SerdeValue::Seq(_) | SerdeValue::Map(_) => {
-            return json_nested_from_serde_value(val, packer, f, descriptors);
+            return nested_datum_from_serde_value(val, packer, f, descriptors, column_type);
         }
         SerdeValue::Char(_) | SerdeValue::Unit | SerdeValue::Newtype(_) => bail!(
             "Unsupported type for Datum from serde_value::Value: {:?}",
@@ -193,100 +195,129 @@ fn default_datum_from_field<'a>(
     }
 }
 
-fn datum_from_serde_proto<'a>(val: &'a ProtoValue) -> Result<Datum<'a>> {
-    match val {
-        ProtoValue::Bool(true) => Ok(Datum::True),
-        ProtoValue::Bool(false) => Ok(Datum::False),
-        ProtoValue::I32(i) => Ok(Datum::Int32(*i)),
-        ProtoValue::I64(i) => Ok(Datum::Int64(*i)),
-        ProtoValue::U32(u) => Ok(Datum::from(Numeric::from(*u))),
-        ProtoValue::U64(u) => Ok(Datum::from(Numeric::from(*u))),
-        ProtoValue::F32(f) => Ok(Datum::Float32((*f).into())),
-        ProtoValue::F64(f) => Ok(Datum::Float64((*f).into())),
-        ProtoValue::String(s) => Ok(Datum::String(s)),
-        ProtoValue::Bytes(b) => Ok(Datum::Bytes(b)),
-        _ => bail!("Unsupported type for Datum from serde_protobuf::Value"),
-    }
-}
-
-fn json_nested_from_serde_value(
+fn nested_datum_from_serde_value(
     val: &SerdeValue,
     packer: &mut Row,
     f: &FieldDescriptor,
     descriptors: &Descriptors,
+    column_type: &ColumnType,
 ) -> Result<()> {
     packer.push(match val {
         SerdeValue::Bool(true) => Datum::True,
         SerdeValue::Bool(false) => Datum::False,
-        SerdeValue::I8(i) => json_number(i)?,
-        SerdeValue::I16(i) => json_number(i)?,
-        SerdeValue::I32(i) => json_number(i)?,
-        SerdeValue::I64(i) => json_number(i)?,
-        SerdeValue::U8(i) => json_number(i)?,
-        SerdeValue::U16(i) => json_number(i)?,
-        SerdeValue::U32(i) => json_number(i)?,
-        SerdeValue::U64(i) => json_number(i)?,
-        SerdeValue::F32(f) => json_number(f)?,
-        SerdeValue::F64(f) => json_number(f)?,
+        SerdeValue::I8(i) => Datum::Int32(*i as i32),
+        SerdeValue::I16(i) => Datum::Int32(*i as i32),
+        SerdeValue::I32(i) => Datum::Int32(*i),
+        SerdeValue::I64(i) => Datum::Int64(*i),
+        SerdeValue::U8(i) => Datum::Int32(*i as i32),
+        SerdeValue::U16(i) => Datum::Int32(*i as i32),
+        SerdeValue::U32(u) => Datum::from(Numeric::from(*u)),
+        SerdeValue::U64(u) => Datum::from(Numeric::from(*u)),
+        SerdeValue::F32(f) => Datum::Float32((*f).into()),
+        SerdeValue::F64(f) => Datum::Float64((*f).into()),
         SerdeValue::String(s) => Datum::String(s),
         SerdeValue::Bytes(_) => {
             bail!("We don't currently support arrays or nested messages with bytes")
         }
         SerdeValue::Seq(s) => {
+            let inner_column_type = if let ColumnType {
+                scalar_type: ScalarType::List { element_type, .. },
+                ..
+            } = column_type
+            {
+                ColumnType {
+                    scalar_type: *element_type.clone(),
+                    nullable: true,
+                }
+            } else {
+                bail!("sequence must be of type list, found {:?}", column_type);
+            };
             return packer.push_list_with(|packer| {
                 for value in s {
-                    json_nested_from_serde_value(&value, packer, f, descriptors)?;
+                    nested_datum_from_serde_value(
+                        &value,
+                        packer,
+                        f,
+                        descriptors,
+                        &inner_column_type,
+                    )?;
                 }
                 Ok(())
             });
         }
         SerdeValue::Option(v) => {
             if let Some(v) = v {
-                return json_nested_from_serde_value(&v, packer, f, descriptors);
+                return nested_datum_from_serde_value(&v, packer, f, descriptors, column_type);
             }
 
             default_datum_from_field_nested(f, descriptors)?
         }
         SerdeValue::Map(m) => {
-            let mut kvs = m.iter().collect::<Vec<_>>();
-            kvs.sort_by(|(k1, _v1), (k2, _v2)| k1.cmp(k2));
-            kvs.dedup_by(|(k1, _v1), (k2, _v2)| k1 == k2);
-            return packer.push_dict_with(|packer| {
-                let nested_message_descriptor = f.field_type(descriptors);
-                for (k, v) in kvs {
-                    match k {
-                        SerdeValue::String(s) => {
+            let nested_message_descriptor = f.field_type(descriptors);
+            let pack_map = |packer: &mut Row,
+                            k: &SerdeValue,
+                            v: &SerdeValue,
+                            typ: &ColumnType,
+                            push_key: bool| {
+                match k {
+                    SerdeValue::String(s) => {
+                        if push_key {
                             packer.push(Datum::String(s.as_str()));
-
-                            let nested_message_descriptor = match nested_message_descriptor {
-                                FieldType::Message(m) => m,
-                                _ => bail!("Nested message is the wrong type"),
-                            };
-
-                            json_nested_from_serde_value(
-                                &v,
-                                packer,
-                                nested_message_descriptor
-                                    .field_by_name(s)
-                                    .expect("Expected this to work"),
-                                descriptors,
-                            )?;
                         }
-                        _ => bail!("Unrecognized value while trying to parse a nested message"),
+
+                        let nested_message_descriptor = match nested_message_descriptor {
+                            FieldType::Message(m) => m,
+                            _ => bail!("Nested message is the wrong type"),
+                        };
+
+                        nested_datum_from_serde_value(
+                            &v,
+                            packer,
+                            nested_message_descriptor
+                                .field_by_name(s)
+                                .expect("nested message to exist"),
+                            descriptors,
+                            typ,
+                        )?;
                     }
+                    _ => bail!("Unrecognized value while trying to parse a nested message"),
                 }
                 Ok(())
-            });
+            };
+
+            match &column_type.scalar_type {
+                ScalarType::Map { value_type, .. } => {
+                    let mut kvs = m.iter().collect::<Vec<_>>();
+                    kvs.sort_by(|(k1, _v1), (k2, _v2)| k1.cmp(k2));
+
+                    return packer.push_dict_with(|packer| {
+                        let typ = ColumnType {
+                            scalar_type: *value_type.clone(),
+                            nullable: true,
+                        };
+                        for (k, v) in kvs {
+                            pack_map(packer, k, v, &typ, true)?;
+                        }
+                        Ok(())
+                    });
+                }
+                ScalarType::Record { fields, .. } => {
+                    return packer.push_list_with(|packer| {
+                        for (n, typ) in fields {
+                            let (k, v) = m
+                                .get_key_value(&SerdeValue::String(n.to_string()))
+                                .expect("key value pair to exist");
+                            pack_map(packer, k, v, &typ, false)?;
+                        }
+                        Ok(())
+                    });
+                }
+                _ => bail!("Unsupported scalar type for map"),
+            }
         }
         _ => bail!("Unsupported types from serde_value"),
     });
     Ok(())
-}
-
-fn json_number<N: ToPrimitive + std::fmt::Display>(i: &N) -> Result<Datum<'static>> {
-    Ok(Datum::Float64(OrderedFloat::from(i.to_f64().ok_or_else(
-        || anyhow!("couldn't convert {} into an f64", i),
-    )?)))
 }
 
 fn default_datum_from_field_nested<'a>(
@@ -329,17 +360,25 @@ fn default_datum_from_field_nested<'a>(
     }
 }
 
-fn datum_from_serde_proto_nested<'a>(val: &'a ProtoValue) -> Result<Datum<'a>> {
+fn datum_from_serde_proto<'a>(val: &'a ProtoValue) -> Result<Datum<'a>> {
     match val {
         ProtoValue::Bool(true) => Ok(Datum::True),
         ProtoValue::Bool(false) => Ok(Datum::False),
-        ProtoValue::I32(i) => json_number(i),
-        ProtoValue::I64(i) => json_number(i),
-        ProtoValue::U32(u) => json_number(u),
-        ProtoValue::U64(u) => json_number(u),
-        ProtoValue::F32(f) => json_number(f),
-        ProtoValue::F64(f) => json_number(f),
+        ProtoValue::I32(i) => Ok(Datum::Int32(*i)),
+        ProtoValue::I64(i) => Ok(Datum::Int64(*i)),
+        ProtoValue::U32(u) => Ok(Datum::from(Numeric::from(*u))),
+        ProtoValue::U64(u) => Ok(Datum::from(Numeric::from(*u))),
+        ProtoValue::F32(f) => Ok(Datum::Float32((*f).into())),
+        ProtoValue::F64(f) => Ok(Datum::Float64((*f).into())),
         ProtoValue::String(s) => Ok(Datum::String(s)),
+        ProtoValue::Bytes(b) => Ok(Datum::Bytes(b)),
         _ => bail!("Unsupported type for Datum from serde_protobuf::Value"),
     }
+}
+
+fn datum_from_serde_proto_nested<'a>(val: &'a ProtoValue) -> Result<Datum<'a>> {
+    if let ProtoValue::Bytes(_) = val {
+        bail!("Nested bytes are not supported");
+    }
+    datum_from_serde_proto(val)
 }