+++ doc/user/content/release-notes.md
@@ -48,6 +48,14 @@ Wrap your release notes at the 80 character mark.
 Put breaking changes before other release notes.
 {{< /comment >}}
 
+{{% version-header v0.9.7 %}}
+- Support the `IS TRUE`, `IS FALSE`, `IS UNKNOWN` operators (and their `NOT`
+  variations). {{% gh 8455 %}}
+- Add support for retention settings on Kafka sinks.
+
+- Support explicit `DROP DATABASE ... (CASCADE | RESTRICT)` statements.  The
+  default behavior remains CASCADE.
+
 {{% version-header v0.9.6 %}}
 
 - Correctly handle TOASTed columns when using PostgreSQL sources. {{% gh 8371 %}}
@@ -85,7 +93,7 @@ Put breaking changes before other release notes.
 - Ignore the trailing newline character of POSIX compliant files instead of
   decoding it as an empty byte row. {{% gh 8142 %}}
 
-- Support `ORDER BY` in aggregate functions.
+- Support `ORDER BY` in [aggregate functions](/sql/functions/#aggregate-func).
 
 - When issuing `COMMIT` or `ROLLBACK` commands outside of an explicit
   transaction, always return a warning. Previously, the warning could be suppressed.